backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Lee Rosenberg <laluk321@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
